notion line formatting attempt with Flutter system.
It was an attempt to replicate feature from notion site
when you type '**text**' and it will automatically apply bold formation to text and delete asterisks.

Problem is with Flutter input system it's hard to replicate.
TextInputConnection produces only TextEditingValues, which contains only raw string and it cannot be overriden.

My path included too much workarounds and dubious solutions.
This commit contains working text editing with bold and italic features (* and ** wrappers).
But for now, I think it will be very tedious to develop this feature further with Flutter input system.